% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKazam.R
\docType{package}
\name{RKazam-package}
\alias{RKazam}
\alias{RKazam-package}
\title{RKazam: Boilerplate for implementing a DBI back end}
\description{
A dummy DBI back end that provides access to a hypothetical
data source named "Kazam", intended to be used as boilerplate for
implementing DBI back ends from scratch.
}
\details{
TBD.
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
