% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Connection.R, R/Result.R
\docType{class}
\name{DBI}
\alias{DBI}
\alias{KazamConnection-class}
\alias{KazamDriver-class}
\alias{KazamResult-class}
\alias{dbBegin,KazamConnection-method}
\alias{dbBind,KazamResult-method}
\alias{dbClearResult,KazamResult-method}
\alias{dbColumnInfo,KazamResult-method}
\alias{dbCommit,KazamConnection-method}
\alias{dbConnect,KazamDriver-method}
\alias{dbDataType,KazamConnection,ANY-method}
\alias{dbDataType,KazamDriver,ANY-method}
\alias{dbDataType,KazamDriver,list-method}
\alias{dbDisconnect,KazamConnection-method}
\alias{dbExistsTable,KazamConnection,character-method}
\alias{dbFetch,KazamResult-method}
\alias{dbGetInfo,KazamConnection-method}
\alias{dbGetInfo,KazamDriver-method}
\alias{dbGetInfo,KazamResult-method}
\alias{dbGetRowCount,KazamResult-method}
\alias{dbGetRowsAffected,KazamResult-method}
\alias{dbGetStatement,KazamResult-method}
\alias{dbHasCompleted,KazamResult-method}
\alias{dbIsValid,KazamConnection-method}
\alias{dbIsValid,KazamDriver-method}
\alias{dbIsValid,KazamResult-method}
\alias{dbListFields,KazamConnection,character-method}
\alias{dbListTables,KazamConnection-method}
\alias{dbQuoteIdentifier,KazamConnection,character-method}
\alias{dbQuoteString,KazamConnection,character-method}
\alias{dbReadTable,KazamConnection,character-method}
\alias{dbRemoveTable,KazamConnection,character-method}
\alias{dbRollback,KazamConnection-method}
\alias{dbSendQuery,KazamConnection,character-method}
\alias{dbWriteTable,KazamConnection,character,data.frame-method}
\alias{show,KazamConnection-method}
\alias{show,KazamDriver-method}
\alias{show,KazamResult-method}
\title{DBI methods}
\usage{
\S4method{show}{KazamDriver}(object)

\S4method{dbConnect}{KazamDriver}(drv, ...)

\S4method{dbDataType}{KazamDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{KazamDriver,list}(dbObj, obj, ...)

\S4method{dbIsValid}{KazamDriver}(dbObj)

\S4method{dbGetInfo}{KazamDriver}(dbObj, ...)

\S4method{show}{KazamConnection}(object)

\S4method{dbDisconnect}{KazamConnection}(conn)

\S4method{dbSendQuery}{KazamConnection,character}(conn, statement, ...)

\S4method{dbDataType}{KazamConnection,ANY}(dbObj, obj, ...)

\S4method{dbQuoteString}{KazamConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{KazamConnection,character}(conn, x, ...)

\S4method{dbWriteTable}{KazamConnection,character,data.frame}(conn, name, value,
  overwrite = FALSE, append = FALSE, ...)

\S4method{dbReadTable}{KazamConnection,character}(conn, name, row.names = NA,
  check.names = TRUE, select.cols = "*")

\S4method{dbListTables}{KazamConnection}(conn)

\S4method{dbExistsTable}{KazamConnection,character}(conn, name)

\S4method{dbListFields}{KazamConnection,character}(conn, name)

\S4method{dbRemoveTable}{KazamConnection,character}(conn, name)

\S4method{dbGetInfo}{KazamConnection}(dbObj, ...)

\S4method{dbIsValid}{KazamConnection}(dbObj)

\S4method{dbBegin}{KazamConnection}(conn)

\S4method{dbCommit}{KazamConnection}(conn)

\S4method{dbRollback}{KazamConnection}(conn)

\S4method{show}{KazamResult}(object)

\S4method{dbClearResult}{KazamResult}(res, ...)

\S4method{dbFetch}{KazamResult}(res, n = -1, ...)

\S4method{dbHasCompleted}{KazamResult}(res, ...)

\S4method{dbGetInfo}{KazamResult}(dbObj, ...)

\S4method{dbIsValid}{KazamResult}(dbObj)

\S4method{dbGetStatement}{KazamResult}(res, ...)

\S4method{dbColumnInfo}{KazamResult}(res, ...)

\S4method{dbGetRowCount}{KazamResult}(res, ...)

\S4method{dbGetRowsAffected}{KazamResult}(res, ...)

\S4method{dbBind}{KazamResult}(res, params, ...)
}
\description{
This package implements pure virtual functions defined in the
\code{DBI} package.
}

