% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Connection.R,
%   R/Result.R, R/dbBegin_KazamConnection.R, R/dbBind_KazamResult.R,
%   R/dbClearResult_KazamResult.R, R/dbColumnInfo_KazamResult.R,
%   R/dbCommit_KazamConnection.R, R/dbConnect_KazamDriver.R,
%   R/dbDataType_KazamConnection.R, R/dbDataType_KazamDriver.R,
%   R/dbDataType_KazamDriver_list.R, R/dbDisconnect_KazamConnection.R,
%   R/dbExistsTable_KazamConnection_character.R,
%   R/dbFetch_KazamResult.R, R/dbGetInfo_KazamConnection.R,
%   R/dbGetInfo_KazamDriver.R, R/dbGetInfo_KazamResult.R,
%   R/dbGetRowCount_KazamResult.R, R/dbGetRowsAffected_KazamResult.R,
%   R/dbGetStatement_KazamResult.R, R/dbHasCompleted_KazamResult.R,
%   R/dbIsValid_KazamConnection.R, R/dbIsValid_KazamDriver.R,
%   R/dbIsValid_KazamResult.R, R/dbListFields_KazamConnection_character.R,
%   R/dbListTables_KazamConnection.R,
%   R/dbQuoteIdentifier_KazamConnection_character.R,
%   R/dbQuoteString_KazamConnection_character.R,
%   R/dbReadTable_KazamConnection_character.R,
%   R/dbRemoveTable_KazamConnection_character.R,
%   R/dbRollback_KazamConnection.R, R/dbSendQuery_KazamConnection_character.R,
%   R/dbSendStatement_KazamConnection_character.R,
%   R/dbWriteTable_KazamConnection_character_data.frame.R,
%   R/show_KazamConnection.R, R/show_KazamDriver.R, R/show_KazamResult.R
\docType{class}
\name{DBI}
\alias{DBI}
\alias{KazamDriver-class}
\alias{KazamConnection-class}
\alias{KazamResult-class}
\alias{dbBegin_KazamConnection}
\alias{dbBegin,KazamConnection-method}
\alias{dbBind_KazamResult}
\alias{dbBind,KazamResult-method}
\alias{dbClearResult_KazamResult}
\alias{dbClearResult,KazamResult-method}
\alias{dbColumnInfo_KazamResult}
\alias{dbColumnInfo,KazamResult-method}
\alias{dbCommit_KazamConnection}
\alias{dbCommit,KazamConnection-method}
\alias{dbConnect_KazamDriver}
\alias{dbConnect,KazamDriver-method}
\alias{dbDataType_KazamConnection}
\alias{dbDataType,KazamConnection,ANY-method}
\alias{dbDataType_KazamDriver}
\alias{dbDataType,KazamDriver,ANY-method}
\alias{dbDataType_KazamDriver_list}
\alias{dbDataType,KazamDriver,list-method}
\alias{dbDisconnect_KazamConnection}
\alias{dbDisconnect,KazamConnection-method}
\alias{dbExistsTable_KazamConnection_character}
\alias{dbExistsTable,KazamConnection,character-method}
\alias{dbFetch_KazamResult}
\alias{dbFetch,KazamResult-method}
\alias{dbGetInfo_KazamConnection}
\alias{dbGetInfo,KazamConnection-method}
\alias{dbGetInfo_KazamDriver}
\alias{dbGetInfo,KazamDriver-method}
\alias{dbGetInfo_KazamResult}
\alias{dbGetInfo,KazamResult-method}
\alias{dbGetRowCount_KazamResult}
\alias{dbGetRowCount,KazamResult-method}
\alias{dbGetRowsAffected_KazamResult}
\alias{dbGetRowsAffected,KazamResult-method}
\alias{dbGetStatement_KazamResult}
\alias{dbGetStatement,KazamResult-method}
\alias{dbHasCompleted_KazamResult}
\alias{dbHasCompleted,KazamResult-method}
\alias{dbIsValid_KazamConnection}
\alias{dbIsValid,KazamConnection-method}
\alias{dbIsValid_KazamDriver}
\alias{dbIsValid,KazamDriver-method}
\alias{dbIsValid_KazamResult}
\alias{dbIsValid,KazamResult-method}
\alias{dbListFields_KazamConnection_character}
\alias{dbListFields,KazamConnection,character-method}
\alias{dbListTables_KazamConnection}
\alias{dbListTables,KazamConnection-method}
\alias{dbQuoteIdentifier_KazamConnection_character}
\alias{dbQuoteIdentifier,KazamConnection,character-method}
\alias{dbQuoteString_KazamConnection_character}
\alias{dbQuoteString,KazamConnection,character-method}
\alias{dbReadTable_KazamConnection_character}
\alias{dbReadTable,KazamConnection,character-method}
\alias{dbRemoveTable_KazamConnection_character}
\alias{dbRemoveTable,KazamConnection,character-method}
\alias{dbRollback_KazamConnection}
\alias{dbRollback,KazamConnection-method}
\alias{dbSendQuery_KazamConnection_character}
\alias{dbSendQuery,KazamConnection,character-method}
\alias{dbSendStatement_KazamConnection_character}
\alias{dbSendStatement,KazamConnection,character-method}
\alias{dbWriteTable_KazamConnection_character_data.frame}
\alias{dbWriteTable,KazamConnection,character,data.frame-method}
\alias{show_KazamConnection}
\alias{show,KazamConnection-method}
\alias{show_KazamDriver}
\alias{show,KazamDriver-method}
\alias{show_KazamResult}
\alias{show,KazamResult-method}
\title{DBI methods}
\usage{
\S4method{dbBegin}{KazamConnection}(conn, ...)

\S4method{dbBind}{KazamResult}(res, params, ...)

\S4method{dbClearResult}{KazamResult}(res, ...)

\S4method{dbColumnInfo}{KazamResult}(res, ...)

\S4method{dbCommit}{KazamConnection}(conn, ...)

\S4method{dbConnect}{KazamDriver}(drv, ...)

\S4method{dbDataType}{KazamConnection,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{KazamDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{KazamDriver,list}(dbObj, obj, ...)

\S4method{dbDisconnect}{KazamConnection}(conn, ...)

\S4method{dbExistsTable}{KazamConnection,character}(conn, name, ...)

\S4method{dbFetch}{KazamResult}(res, n = -1, ...)

\S4method{dbGetInfo}{KazamConnection}(dbObj, ...)

\S4method{dbGetInfo}{KazamDriver}(dbObj, ...)

\S4method{dbGetInfo}{KazamResult}(dbObj, ...)

\S4method{dbGetRowCount}{KazamResult}(res, ...)

\S4method{dbGetRowsAffected}{KazamResult}(res, ...)

\S4method{dbGetStatement}{KazamResult}(res, ...)

\S4method{dbHasCompleted}{KazamResult}(res, ...)

\S4method{dbIsValid}{KazamConnection}(dbObj, ...)

\S4method{dbIsValid}{KazamDriver}(dbObj, ...)

\S4method{dbIsValid}{KazamResult}(dbObj, ...)

\S4method{dbListFields}{KazamConnection,character}(conn, name, ...)

\S4method{dbListTables}{KazamConnection}(conn, ...)

\S4method{dbQuoteIdentifier}{KazamConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{KazamConnection,character}(conn, x, ...)

\S4method{dbReadTable}{KazamConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{KazamConnection,character}(conn, name, ...)

\S4method{dbRollback}{KazamConnection}(conn, ...)

\S4method{dbSendQuery}{KazamConnection,character}(conn, statement, ...)

\S4method{dbSendStatement}{KazamConnection,character}(conn, statement, ...)

\S4method{dbWriteTable}{KazamConnection,character,data.frame}(conn, name, value, overwrite = FALSE, append = FALSE, ...)

\S4method{show}{KazamConnection}(object)

\S4method{show}{KazamDriver}(object)

\S4method{show}{KazamResult}(object)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{...}{Other parameters passed on to methods.}

\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{params}{A list of bindings, named or unnamed.}

\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{dbObj}{A object inheriting from \linkS4class{DBIDriver}
or \linkS4class{DBIConnection}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{x}{A character vector, \link[DBI]{SQL} or \link[DBI]{Id} object to quote as identifier.}

\item{statement}{a character string containing SQL.}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{overwrite}{Allow overwriting the destination table. Cannot be
\code{TRUE} if \code{append} is also \code{TRUE}.}

\item{append}{Allow appending to the destination table. Cannot be
\code{TRUE} if \code{overwrite} is also \code{TRUE}.}

\item{object}{Any R object}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package.
}
